+++ lib/grape/middleware/versioner/header.rb
@@ -1,4 +1,5 @@
 require 'grape/middleware/base'
+require 'grape/middleware/versioner/parse_media_type_patch'
 
 module Grape
   module Middleware